+++ src/main/java/com/studyolle/settings/validator/NickNameFormValidator.java
@@ -1,7 +1,8 @@
-package com.studyolle.settings;
+package com.studyolle.settings.validator;
 
 import com.studyolle.account.AccountRepository;
 import com.studyolle.domain.Account;
+import com.studyolle.settings.form.NicknameForm;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
